requirements.txt
PyPDF2==3.0.1
pikepdf==8.3.0
reportlab==4.0.4
python-dotenv==1.0.0
tqdm==4.66.1
PyQt5==5.15.10